requirements.txt
torch>=2.2
torchvision>=0.17
numpy
scipy
scikit-learn
matplotlib
tqdm
pyyaml
pandas
openml
